lib folder moved to src/lib
