+++ src/lib/fs-helper.js
