+++ test/common.test.js
@@ -1,5 +1,5 @@
-const { createConfig } = require("../lib/common");
-jest.mock('../lib/action-utils', () => ({
+const { createConfig } = require("../src/lib/common");
+jest.mock('../src/lib/action-utils', () => ({
   getParentDependencies: () => { return ['lienzo-core', 'lienzo-test', 'drools']; },
   getChildDependencies: () => { return ['lienzo-core', 'lienzo-test', 'drools-jbpm']; }
 }));
